feat(fe):  add group admin problem list
